Refactor: move files to components libs or hooks folders on specific path folders
