+++ src/app/haku/[oid]/hakukohde/[hakukohde]/hakijaryhmat/page.tsx
@@ -7,10 +7,10 @@ import { Box, CircularProgress } from '@mui/material';
 import { useSuspenseQueries } from '@tanstack/react-query';
 import { getHakijaryhmat } from '@/app/lib/valintalaskenta-service';
 import { isEmpty } from '@/app/lib/common';
-import { HakijaryhmaContent } from './hakijaryhma-content';
+import { HakijaryhmaContent } from './components/hakijaryhma-content';
 import { PageSizeSelector } from '@/app/components/table/page-size-selector';
 import { useHakijaryhmatSearchParams } from '@/app/hooks/useHakijaryhmatSearch';
-import { HakijaryhmatControls } from './hakijaryhmat-controls';
+import { HakijaryhmatControls } from './components/hakijaryhmat-controls';
 import { NoResults } from '@/app/components/no-results';
 
 type HakijaryhmatContentParams = {
