+++ src/app/haku/[oid]/hakukohde/[hakukohde]/perustiedot/page.tsx
@@ -2,9 +2,9 @@
 
 import { getSijoittelunTulokset } from '@/app/lib/valinta-tulos-service';
 import { TabContainer } from '../tab-container';
-import BasicInfo from './basic-info';
+import BasicInfo from './components/basic-info';
 import { useSuspenseQueries } from '@tanstack/react-query';
-import { ValintatapajonotTable } from './valintatapajonot-table';
+import { ValintatapajonotTable } from './components/valintatapajonot-table';
 import { getHaku } from '@/app/lib/kouta';
 import { useTranslations } from '@/app/hooks/useTranslations';
 import { QuerySuspenseBoundary } from '@/app/components/query-suspense-boundary';
