+++ src/app/haku/[oid]/hakukohde/[hakukohde]/pistesyotto/components/table/koe-readonly-cell.tsx
