+++ src/app/haku/[oid]/hakukohde/[hakukohde]/pistesyotto/components/table/pistesyotto-table.tsx
