+++ src/app/haku/[oid]/hakukohde/[hakukohde]/pistesyotto/hooks/usePisteSyottoSearch.ts
@@ -3,15 +3,18 @@ import { useEffect, useMemo } from 'react';
 import { useDebounce } from '@/app/hooks/useDebounce';
 import { parseAsBoolean, parseAsInteger, useQueryState } from 'nuqs';
 import { useHasChanged } from '@/app/hooks/useHasChanged';
-import { byProp, getSortParts } from '../components/table/table-utils';
+import {
+  byProp,
+  getSortParts,
+} from '../../../../../../components/table/table-utils';
 import {
   DEFAULT_PAGE_SIZE,
   HAKU_SEARCH_PHRASE_DEBOUNCE_DELAY,
 } from '@/app/lib/constants';
-import { useTranslations } from './useTranslations';
-import { HakukohteenPistetiedot } from '../lib/types/laskenta-types';
-import { hakemusFilter } from './filters';
-import { DEFAULT_NUQS_OPTIONS } from './common';
+import { useTranslations } from '../../../../../../hooks/useTranslations';
+import { HakukohteenPistetiedot } from '../../../../../../lib/types/laskenta-types';
+import { hakemusFilter } from '../../../../../../hooks/filters';
+import { DEFAULT_NUQS_OPTIONS } from '../../../../../../hooks/common';
 
 export const usePisteSyottoSearchParams = () => {
   const [searchPhrase, setSearchPhrase] = useQueryState(
