+++ src/app/haku/[oid]/hakukohde/[hakukohde]/pistesyotto/lib/pistesyotto-utils.ts
