+++ src/app/haku/[oid]/hakukohde/[hakukohde]/pistesyotto/page.tsx
@@ -5,9 +5,9 @@ import { QuerySuspenseBoundary } from '@/app/components/query-suspense-boundary'
 import { ClientSpinner } from '@/app/components/client-spinner';
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { getPisteetForHakukohde } from '@/app/lib/valintalaskentakoostepalvelu';
-import { PisteSyottoControls } from './pistesyotto-controls';
+import { PisteSyottoControls } from './components/pistesyotto-controls';
 import { Box } from '@mui/material';
-import { PisteSyottoForm } from './pistesyotto-form';
+import { PisteSyottoForm } from './components/pistesyotto-form';
 import { useTranslations } from '@/app/hooks/useTranslations';
 import { isEmpty } from '@/app/lib/common';
 import { NoResults } from '@/app/components/no-results';
