+++ src/app/haku/[oid]/hakukohde/[hakukohde]/valinnan-hallinta/lib/valinnan-hallinta-types.ts
