+++ src/app/haku/[oid]/hakukohde/[hakukohde]/valintalaskennan-tulos/components/laskettu-valintatapajono-table.tsx
